refactored logic out of activities to activityview | route setup to use activityview component.
